+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/base/Repository.java
@@ -1,6 +1,6 @@
 package com.gabrielosorio.gestor_inteligente.repository.base;
 
-import com.gabrielosorio.gestor_inteligente.repository.specification.Specification;
+import com.gabrielosorio.gestor_inteligente.repository.specification.base.Specification;
 import com.gabrielosorio.gestor_inteligente.repository.strategy.base.BatchInsertable;
 import com.gabrielosorio.gestor_inteligente.repository.strategy.base.RepositoryStrategy;
 import com.gabrielosorio.gestor_inteligente.repository.strategy.base.TransactionalStrategy;
