+++ src/main/java/com/gabrielosorio/gestor_inteligente/repository/specification/SaleProductBySaleIdSpec.java
@@ -2,6 +2,8 @@ package com.gabrielosorio.gestor_inteligente.repository.specification;
 import com.gabrielosorio.gestor_inteligente.config.DBScheme;
 import com.gabrielosorio.gestor_inteligente.config.QueryLoader;
 import com.gabrielosorio.gestor_inteligente.model.SaleProduct;
+import com.gabrielosorio.gestor_inteligente.repository.specification.base.Specification;
+
 import java.util.List;
 
 public class SaleProductBySaleIdSpec implements Specification<SaleProduct> {
